requirements.txt
asyncio==3.4.3
dnspython==2.7.0
grpcio==1.68.0
grpcio-tools==1.68.0
motor==3.6.0
protobuf==5.28.3
py-dotenv==0.1
pymongo==4.9.2
python-dotenv==1.0.1
setuptools==75.6.0
